:recycle: Refactor: organize files and apply feedbacks -211125
