+++ client/src/screens/drawer-select/TeamSelect.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import styled from 'styled-components/native';
-import BottomDrawer from '../components/drawer/BottomDrawer';
-import { colors } from '../theme/colors';
-import { Bold, Light } from '../theme/fonts';
+import BottomDrawer from '../../components/drawer/BottomDrawer';
+import { colors } from '../../theme/colors';
+import { Bold, Light } from '../../theme/fonts';
 import { MaterialIcons } from '@expo/vector-icons';
 import { NavigationProp, useNavigation } from '@react-navigation/native';
-import { RootStackParamList } from '../navigation/RootStackParamList';
+import { RootStackParamList } from '../../navigation/RootStackParamList';
+import CaptainMark from '../../components/mark/CaptainMark';
 
 const TitleContainer = styled.View`
   margin: 15px 0px;
@@ -34,14 +35,6 @@ const Check = styled.View`
   margin-left: 20px;
 `;
 
-export const CaptainMark = styled.View`
-  width: 28px;
-  height: 20px;
-  justify-content: center;
-  align-items: center;
-  background-color: ${colors.yellow};
-`;
-
 export default function TeamSelect() {
   const navigation = useNavigation<NavigationProp<RootStackParamList>>();
   return (
@@ -52,9 +45,7 @@ export default function TeamSelect() {
       <TeamContainer>
         <Team>FC 프론트</Team>
         <IconContainer>
-          <CaptainMark>
-            <Bold>C</Bold>
-          </CaptainMark>
+          <CaptainMark />
           <Check>
             <MaterialIcons name="check" size={20} color={colors.blue} />
           </Check>
@@ -66,7 +57,7 @@ export default function TeamSelect() {
       <TeamContainer>
         <Team>FC 살쾡이</Team>
       </TeamContainer>
-      <TeamContainer onPress={() => navigation.navigate('CreateTeam')}>
+      <TeamContainer onPress={() => navigation.navigate('Home')}>
         <NewTeam>+ 새로운 소속팀</NewTeam>
       </TeamContainer>
     </BottomDrawer>
