--- client/src/theme/index.ts
@@ -1,14 +1,0 @@
-export const colors = {
-  white: '#FFFFF',
-  darkGray: '#3C3C3C',
-  gray: '868686',
-  lightGray: 'DDDDDD',
-  ivory: 'F5F5F5',
-  blue: '#006FAD',
-  red: '#AD0000',
-  yellow: '#FEE500',
-};
-
-export const fonts = {
-  primary: 'Spoqa Han Sans Neo',
-};
